# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Delivery Microservice",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/webhooks/orders": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Actualiza una orden",
                "operationId": "6a68d11e2502fc73a648e9e6df1d646e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event",
                                    "data"
                                ],
                                "properties": {
                                    "event": {
                                        "type": "string",
                                        "example": "update_cooking_status"
                                    },
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "order_code"
                                            ],
                                            "properties": {
                                                "order_code": {
                                                    "type": "string",
                                                    "example": "2302faca-7f66-4078-86d4-abb0ab54b675"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "2302faca-7f66-4078-86d4-abb0ab54b675"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}